Updated Gradle version and project dependencies
Updated the project's Gradle version to 8.3 from 7.5 and updated the project version from 1.0.0 to 1.0.1 for development progression. The kodein-internal-gradle-settings dependency was also updated for the same reason. Removed the Kodein JVM version specification as the build will be done on JDK 17 but will maintain compatibility with older versions of the JDK.